Move singleton to utils
