+++ src/utils/singleton.py
@@ -1,0 +1,14 @@
+class Singleton(type):
+    _instances = {}
+
+    def __call__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            instance = super().__call__(*args, **kwargs)
+            cls._instances[cls] = instance
+        return cls._instances[cls]
+
+    def instance(cls):
+        if cls not in cls._instances:
+            return None
+
+        return cls._instances[cls]
